because of a bug in handling messages with sequence numbers, it was possible for messages to be stored in the database with missing sequence numbers.  I.e. numbers 1, 2, and 5 could be stored leaving 3 and 4 missing.
this results in the problem that the message with sequence number 5 is
believed to be a message with sequence number 3.  resulting in an
inconsistent database and an inability to correctly handle missing
sequence messages and incoming messages with specific sequence
numbers.
            
we will 'solve' this by removing all messages after a 'gap' occurred
in the sequence numbers.  In our example it will result in the message
with sequence number 5 to be removed.
            
we choose not to call any undo methods because both the timeline and
the votes can handle the resulting multiple calls to the undo
callback.


git-svn-id: https://svn.tribler.org/dispersy/branches/20120612-27046-mainbranch@28701 001aeff7-3401-0410-a489-f7902fc005dd